Version 1.2.2
Remove slime splitting as a natural-like spawn method because it effectively
converts custom spawners (slime or magma cube) into sources of large numbers
of mobs eligible for head and EoF drops.

This does mean that only the primary, natural spawn of slimes and magma
cubes is eligible for custom drops, so elevate drop probabilities for these.